[refacotr] Move formatDateToYYYYMMDD to util
