+++ src/main/frontend/src/util/DateUtil.js
@@ -44,3 +44,17 @@ export const parseDate = (dateString) => {
 
   return new Date(...args);
 };
+
+export const formatDateToYYYYMMDD = (date) => {
+  if (!(date instanceof Date)) {
+    date = new Date(date);
+  }
+  let month = "" + (date.getMonth() + 1);
+  let day = "" + date.getDate();
+  let year = date.getFullYear();
+
+  if (month.length < 2) month = "0" + month;
+  if (day.length < 2) day = "0" + day;
+
+  return [year, month, day].join("");
+};
